Factor out ContentFrame from local-page, use React.
