+++ src/local-page/local-page.css
@@ -4,7 +4,11 @@ body {
     height: 100%;
     width: 100%;
     box-sizing: border-box;
+}
 
+#rootContainer {
+    width: 100%;
+    height: 100%;
     display: flex;
     flex-direction: column;
 }
