+++ src/local-page/local-page.jsx
@@ -8,19 +8,9 @@ import { getPage } from 'src/search/find-pages'
 import { getTimestamp } from 'src/activity-logger'
 import shortUrl from 'src/util/short-url'
 import niceTime from 'src/util/nice-time'
-import syncLocationHashes from 'src/util/sync-location-hashes'
 
+import ContentFrame from './ContentFrame'
 
-function fixChromiumInjectedStylesheet(document) {
-    // Pragmatic workaround for Chromium, which appears to inject two style
-    // rules into extension pages (with font-size: 75%, for some reason?).
-    const styleEl = document.createElement('style')
-    styleEl.innerHTML = `body {
-        font-size: inherit;
-        font-family: inherit;
-    }`
-    document.head.insertAdjacentElement('afterbegin', styleEl)
-}
 
 async function saveAs({page}) {
     const pageId = page._id
@@ -84,41 +74,12 @@ async function showPage(pageId) {
         </div>
     )
     ReactDOM.render(
-        bar,
+        <div id='rootContainer'>
+            {bar}
+            <ContentFrame html={html} />
+        </div>,
         document.getElementById('app')
     )
-
-    // Show the page in the iframe.
-    const iframe = document.createElement('iframe')
-    iframe.setAttribute('sandbox', 'allow-same-origin allow-top-navigation allow-scripts')
-    iframe.setAttribute('id', 'page')
-    iframe.setAttribute('seamless', 'seamless')
-    iframe.srcdoc = html
-    // XXX The DOMContentLoaded event would be better, but how to listen to that?
-    iframe.onload = () => {
-        const doc = iframe.contentDocument
-
-        // Ensure a head element exists.
-        if (!doc.head) {
-            const head = doc.createElement('head')
-            doc.documentElement.insertAdjacentElement('afterbegin', head)
-        }
-
-        // Make links open in the whole tab, not inside the iframe
-        const baseEl = doc.createElement('base')
-        baseEl.setAttribute('target', '_parent')
-        doc.head.insertAdjacentElement('afterbegin', baseEl)
-
-        // Workaround required for Chromium.
-        fixChromiumInjectedStylesheet(doc)
-
-        // Focus on the page so it receives e.g. keyboard input
-        iframe.contentWindow.focus()
-
-        // Keep the iframe's location #hash in sync with that of the window.
-        syncLocationHashes([window, iframe.contentWindow], {initial: window})
-    }
-    document.body.appendChild(iframe)
 }
 
 // Read pageId from location: ?page=pageId
